+++ packages/client-flows/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [typescript v1.0.0-alpha.2] packages/client-flows
 ### Releases
 - [NPM v1.0.0-alpha.2] https://www.npmjs.com/package/@formance/sdk-flows/v/1.0.0-alpha.2 - packages/client-flows
+
+## 2024-07-11 00:12:23
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.331.0 (2.365.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.3] packages/client-flows
+### Releases
+- [NPM v1.0.0-alpha.3] https://www.npmjs.com/package/@formance/sdk-flows/v/1.0.0-alpha.3 - packages/client-flows